Arreglo de bug en el método de busqueda por parametros y cambios menores en clases

+++ src/main/java/com/tienda/cincomenos/domain/producto/dto/DatosActualizarProducto.java
@@ -1,5 +1,6 @@
 package com.tienda.cincomenos.domain.producto.dto;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 import com.tienda.cincomenos.domain.producto.CategoriaProducto;
@@ -12,7 +13,7 @@ public record DatosActualizarProducto(
     String nombre, 
     String descripcion, 
     String marca, 
-    Double precio,
+    BigDecimal precio,
     @NotNull
     CategoriaProducto categoria,
     Map<String, String> atributosDeSubclases) {
